Complete redesign
I have reworked the design of the components streamlining the code and at the same time allowing easier integration of other algorithms based on similar ideas. This really is step one in the modernization and the next step would be the use of interfaces. I haven't changed the registration palette but I strongly advise that to be moved from "Samples" into its own tab.